requirements.txt
# 核心量子计算库
qiskit>=0.40.0
qiskit-aer>=0.12.0

# 数学和科学计算
numpy>=1.21.0
scipy>=1.7.0

# 可视化
matplotlib>=3.5.0

# 图像处理（可选）
opencv-python>=4.5.0
Pillow>=8.0.0

# 安装说明：
# 1. 如果使用新版本Qiskit (>=1.0)，请运行：
#    pip install qiskit qiskit-aer
# 
# 2. 如果使用旧版本Qiskit (<1.0)，请运行：
#    pip install qiskit[all]
#
# 3. 最小安装（仅核心功能）：
#    pip install qiskit numpy matplotlib